Updated values and added some consistency across the models :§

+++ pokepi/src/models/location.rs
@@ -2,8 +2,8 @@ use serde::Serialize;
 
 #[derive(Debug, Serialize)]
 pub struct Location {
-    pub location_id: u16,
+    pub location_id: u32,
     pub location_name: String,
     pub location_identifier: String,
-    pub region_id: i16,
+    pub region_id: u16,
 }
