+++ pokepi/src/models/pokemon_form_ability.rs
@@ -2,8 +2,8 @@ use serde::Serialize;
 
 #[derive(Debug, Serialize)]
 pub struct PokemonFormAbility {
-    pub pokemon_form_game_id: i16,
-    pub ability_id: i16,
-    pub slot: i8,
+    pub pokemon_form_game_id: u32,
+    pub ability_id: u16,
+    pub slot: u8,
     pub is_hidden: bool,
 }
