+++ pokepi/src/models/type_matchup.rs
@@ -2,7 +2,7 @@ use serde::Serialize;
 
 #[derive(Debug, Serialize)]
 pub struct TypeMatchup {
-    pub attacking_type_id: i8,
-    pub defending_type_id: i8,
-    pub damage_factor: f64,
+    pub attacking_type_id: u8,
+    pub defending_type_id: u8,
+    pub damage_factor: f32,
 }
